+++ resources/views/fields/checkbox.blade.php
@@ -10,11 +10,7 @@
 
     @include('livewire-forms::fields.label')
 
-    @includeWhen(
-        $field->tooltip,
-        'livewire-forms::components.tooltip',
-        ['text' => $field->tooltip]
-    )
+    @include('livewire-forms::fields.gdpr')
 
     @include('livewire-forms::fields.error')
 </div>