+++ resources/views/fields/image.blade.php
@@ -2,14 +2,15 @@
     @include('livewire-forms::fields.label')
 
     @if($field->getValue())
-    <img
-        class="{{ $field->class ?? 'img-fluid' }}"
-        src="{{ isset($files_[$field->getName()]) ? $files_[$field->getName()]->temporaryUrl() : $field->getValue() }}"
-        alt="{{ $field->altText ?? optional($field->getValue())->name  }}"
-        id="{{ $field->getName() }}"
-        @if ($field->dusk) dusk={{ $field->dusk }} @endif
-    >
+        <img
+            class="{{ $field->class ?? 'img-fluid' }}"
+            src="{{ isset($files_[$field->getName()]) ? $files_[$field->getName()]->temporaryUrl() : $field->getValue() }}"
+            alt="{{ $field->altText ?? optional($field->getValue())->name  }}"
+            id="{{ $field->getName() }}"
+            @if ($field->dusk) dusk={{ $field->dusk }} @endif
+        >
     @endif
+
     <input
         type="file"
         class="{{ $field->class }}"
@@ -20,11 +21,7 @@
         @if ($field->dusk) dusk={{ $field->dusk }} @endif
     >
 
-    @includeWhen(
-        $field->tooltip,
-        'livewire-forms::components.tooltip',
-        ['text' => $field->tooltip]
-    )
+    @include('livewire-forms::fields.gdpr')
 
     @include('livewire-forms::fields.error')
 </div>