+++ resources/views/fields/select.blade.php
@@ -1,6 +1,5 @@
 <div class="{{ $field->divClass ?? config('livewire-forms.defaults.divClass') }}">
     @include('livewire-forms::fields.label')
-
     <select
         @include('livewire-forms::fields.binding')
         id="{{ $field->getName() }}"
@@ -11,22 +10,13 @@
     >
         <option value="">{{ __('form.select an option') }}</option>
         @foreach ($field->options as $key => $value)
-            <option
-                value="{{ $field->useValueAsKeys ? $value : $key  }}"
-                @if($field->getValue() == $key)
-                    selected
-                @endif
-            >
+            <option value="{{ $field->useValueAsKeys ? $value : $key }}">
                 {{ $value }}
             </option>
         @endforeach
     </select>
 
-    @includeWhen(
-        $field->tooltip,
-        'livewire-forms::components.tooltip',
-        ['text' => $field->tooltip]
-    )
+    @include('livewire-forms::fields.gdpr')
 
     @include('livewire-forms::fields.error')
 </div>