+++ resources/views/fields/textarea.blade.php
@@ -10,11 +10,7 @@
         @if ($field->dusk) dusk={{ $field->dusk }} @endif
     ></textarea>
 
-    @includeWhen(
-        $field->tooltip,
-        'livewire-forms::components.tooltip',
-        ['text' => $field->tooltip]
-    )
+    @include('livewire-forms::fields.gdpr')
 
     @include('livewire-forms::fields.error')
 </div>