+++ src/Fields/Spacer.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Codedor\LivewireForms\Fields;
+
+class Spacer extends Field
+{
+    public $component = 'livewire-forms::fields.spacer';
+
+    public $isField = false;
+}